# swagger.yaml
openapi: 3.0.3
info:
  title: PLM Costing Module API
  description: |
    API for PLM Costing Module data manipulation and integration.
    
    This API receives XML data from PLM system, extracts relevant information (ModuleId and other properties),
    performs necessary data manipulations, and will integrate with PLM's patch APIs to write back to the system.
    
    **Base URL:** https://costingcalculations-4ece4d469e2a.herokuapp.com
    
  version: 1.0.0
  contact:
    name: API Support
    email: kaan.karaca@kblplus.com
  license:
    name: ISC
servers:
  - url: https://costingcalculations-4ece4d469e2a.herokuapp.com
    description: Production Server (Heroku)
  - url: http://localhost:3000
    description: Local Development Server
...
paths:
...
  /api/workflow/process:
    post:
      tags:
        - Costing
      summary: Process PLM Workflow XML (Main Endpoint)
      description: |
        **Main endpoint for all PLM workflow processing.**
        
        This endpoint receives ProcessWorkflow XML from PLM system, determines the workflow type
        based on WorkflowDefinitionCode, and routes to appropriate handler.
        
        **Supported Workflows:**
        - `UPDATED_STYLE_OVERVIEW`: Routes to overview-to-costing flow
          - Fetches style costing data from PLM
          - Performs costing calculations using decision table
          - PATCH results back to PLM (StyleCostingSupplierValue + StyleExtendedFieldValues)
        
        **Workflow Routing:**
        The API automatically routes based on WorkflowDefinitionCode in the XML.
        Unknown workflow types return 200 with "route not found" message.
        
        **Process Flow:**
        1. Extract WorkflowDefinitionCode and ModuleId from XML
        2. Route to appropriate workflow handler
        3. Fetch data from PLM GET API
        4. Perform calculations using decision table
        5. PATCH results back to PLM
        6. Return success response with calculated data
        
      operationId: processWorkflow
      requestBody:
        description: Workflow data from ION (JSON format)
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - workflowdefination
                - moduleId
              properties:
                workflowdefination:
                  type: string
                  description: Workflow type code
                  enum:
                    - UPDATED_STYLE_OVERVIEW
                    - UPDATED_STYLE_BOO
                    - UPDATED_STYLE_BOM
                moduleId:
                  type: string
                  description: Style ID from PLM
              example:
                workflowdefination: "UPDATED_STYLE_OVERVIEW"
                moduleId: "158"
          text/xml:
            schema:
              type: string
              description: XML format (legacy support)
              example: |
                <?xml version="1.0" encoding="utf-8"?>
                <ProcessWorkflow xmlns="http://schema.infor.com/InforOAGIS/2">
                  <ApplicationArea>
                    <Sender>
                      <LogicalID>lid://infor.fashionplm.fplm</LogicalID>
                      <ComponentID>FPLM</ComponentID>
                    </Sender>
                    <CreationDateTime>2025-12-14T17:57:43Z</CreationDateTime>
                  </ApplicationArea>
                  <DataArea>
                    <Process>
                      <TenantID>ATJZAMEWEF5P4SNV_TST</TenantID>
                    </Process>
                    <Workflow>
                      <Status><Code>Initial</Code></Status>
                      <WorkflowDefinitionCode>UPDATED_STYLE_OVERVIEW</WorkflowDefinitionCode>
                      <Property>
                        <NameValue name="ModuleId" type="StringType">9457</NameValue>
                      </Property>
                      <Property>
                        <NameValue name="ModuleCode" type="StringType">I00000120101</NameValue>
                      </Property>
                      <Property>
                        <NameValue name="ModuleName" type="StringType">BEL CANTASI</NameValue>
                      </Property>
                    </Workflow>
                  </DataArea>
                </ProcessWorkflow>
      responses:
        '200':
          description: Successfully processed workflow
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/WorkflowSuccessResponse'
                  - $ref: '#/components/schemas/WorkflowUnknownResponse'
              examples:
                successfulCosting:
                  summary: Successful Costing Calculation
                  value:
                    success: true
                    workflow: "OVERVIEW_TO_COSTING"
                    styleId: "9457"
                    calculatedData:
                      StyleId: 9457
                      StyleCode: "I00000120101"
                      BrandId: 4
                      SubCategoryId: 41
                      UserDefinedField5Id: 1
                      Cluster: "016-E"
                      SPSFValue: 4699
                      MUValue: 4.8
                      KHDFValue: 0
                      ALMUSDValue: 18.54
                      ALMTRYValue: 889.96
                      SPSFId: 1295
                      MUId: 1335
                    patchResults:
                      styleCostingSupplierValue: {}
                      styleExtendedFieldValues: {}
                    message: "Costing calculation and PATCH completed successfully"
                    timestamp: "2025-12-15T10:00:00.000Z"
                unknownWorkflow:
                  summary: Unknown Workflow Type
                  value:
                    success: true
                    message: "Workflow received but route not found"
                    workflowType: "SOME_OTHER_WORKFLOW"
                    moduleId: "9457"
                    timestamp: "2025-12-15T10:00:00.000Z"
        '400':
          description: Bad Request - No XML data or invalid format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "No XML data received"
                message: "Request body is empty"
        '404':
          description: Required data not found (ModuleId, WorkflowDefinitionCode, or Style data)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                noModuleId:
                  summary: ModuleId not found
                  value:
                    error: "ModuleId not found"
                    message: "ModuleId property not found in the XML"
                noWorkflowCode:
                  summary: WorkflowDefinitionCode not found
                  value:
                    error: "WorkflowDefinitionCode not found"
                    message: "WorkflowDefinitionCode not found in the XML"
                noStyle:
                  summary: Style not found in PLM
                  value:
                    error: "Style not found"
                    message: "No style data found for StyleId: 9457"
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "Internal server error"
                message: "XML parsing failed: Unexpected token"
  /api/workflow/get-cost-element-values:
    post:
      tags:
        - Costing
      summary: Get Cost Element Values from PLM
      description: |
        **Retrieves specific cost element values from PLM for the main supplier.**
        
        This endpoint fetches cost element values (RHDF, GKUR, TKMS, TAST, TISC, TTRM, TISL, TDGR)
        for a specific style. It automatically selects the main supplier based on:
        - IsMainVersion = true
        - IsActive = true
        - Highest Id among matching suppliers
        
        **Process Flow:**
        1. Fetch style costing data from PLM using provided filter
        2. Find main supplier (IsMainVersion=true, IsActive=true, highest Id)
        3. Extract cost element values for selected supplier
        4. Return values as JSON object
        
      operationId: getCostElementValues
      requestBody:
        description: Filter to identify the style
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - filter
              properties:
                filter:
                  type: string
                  description: OData filter expression (e.g., "StyleId eq 10596")
                  example: "StyleId eq 10596"
            example:
              filter: "StyleId eq 10596"
      responses:
        '200':
          description: Successfully retrieved cost element values
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CostElementValuesResponse'
              example:
                success: true
                styleId: 10596
                styleCode: "IS1230034100"
                selectedSupplierId: 178
                values:
                  RHDF: 0.000000
                  GKUR: 0.000000
                  TKMS: 0.000000
                  TAST: 0.000000
                  TISC: 0.000000
                  TTRM: 0.000000
                  TISL: 0.000000
                  TDGR: 0.000000
                  TCOST: 0.000000
                  RMU: 0.000000
                extendedFieldIds:
                  AlimFiyati_USD: 196065
                  SegmentPSF: 196066
                  KumasHedefMaliyet: 196067
                  AlimFiyat_TRY: 196068
                  AlimTarget_USD: 196069
                  AlimTarget_USD_105: 196070
                  TKMS: 196071
                  TAST: 196072
                  TISC: 196073
                  TTRM: 196074
                  TISL: 196075
                  TDGR: 196076
                timestamp: "2025-01-21T12:00:00.000Z"
        '400':
          description: Bad Request - Missing filter
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                errorCode: "MISSING_FILTER"
                error: "Filter is required"
                message: "Please provide a filter parameter (e.g., 'StyleId eq 10596')"
                timestamp: "2025-01-21T12:00:00.000Z"
        '404':
          description: Style or costing data not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                styleNotFound:
                  summary: Style not found
                  value:
                    success: false
                    errorCode: "STYLE_NOT_FOUND"
                    error: "Style not found"
                    message: "No style found matching the provided filter"
                    filter: "StyleId eq 99999"
                    timestamp: "2025-01-21T12:00:00.000Z"
                noCostingData:
                  summary: No costing data
                  value:
                    success: false
                    errorCode: "NO_COSTING_DATA"
                    error: "No costing data found for this style"
                    message: "Style exists but has no costing data"
                    styleId: 10596
                    timestamp: "2025-01-21T12:00:00.000Z"
                noMainSupplier:
                  summary: No main supplier found
                  value:
                    success: false
                    errorCode: "NO_MAIN_SUPPLIER"
                    error: "No main active supplier found"
                    message: "No supplier with IsMainVersion=true and IsActive=true"
                    styleId: 10596
                    timestamp: "2025-01-21T12:00:00.000Z"
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                errorCode: "INTERNAL_ERROR"
                error: "Database connection failed"
                message: "Error retrieving cost element values"
                timestamp: "2025-01-21T12:00:00.000Z"
components:
  schemas:
...
    WorkflowSuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          description: Indicates successful processing
          example: true
        workflow:
          type: string
          description: Workflow type that was processed
          example: "OVERVIEW_TO_COSTING"
        styleId:
          type: string
          description: Style ID (ModuleId) from XML
          example: "9457"
        calculatedData:
          type: object
          description: Calculated costing data with decision table values
          properties:
            StyleId:
              type: integer
            StyleCode:
              type: string
            BrandId:
              type: integer
            SubCategoryId:
              type: integer
            UserDefinedField5Id:
              type: integer
            Cluster:
              type: string
            SPSFValue:
              type: number
            MUValue:
              type: number
            KHDFValue:
              type: number
            ALMUSDValue:
              type: number
            ALMTRYValue:
              type: number
            SPSFId:
              type: integer
            MUId:
              type: integer
        patchResults:
          type: object
          description: Results from PATCH operations to PLM
        message:
          type: string
          example: "Costing calculation and PATCH completed successfully"
        timestamp:
          type: string
          format: date-time
          example: "2025-12-15T10:00:00.000Z"
      required:
        - success
        - workflow
        - styleId
        - calculatedData
        - message
        - timestamp
    WorkflowUnknownResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "Workflow received but route not found"
        workflowType:
          type: string
          example: "UNKNOWN_WORKFLOW"
        moduleId:
          type: string
          example: "9457"
        timestamp:
          type: string
          format: date-time
      required:
        - success
        - message
        - workflowType
        - moduleId
        - timestamp
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error type/category
          example: "No XML data received"
        message:
          type: string
          description: Detailed error message
          example: "Request body is empty"
      required:
        - error
        - message
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT Bearer token authentication (for future implementation).
        Currently, the API does not require authentication.
security: []